Feature/auth module

+++ src/cv/cv.service.ts
@@ -9,6 +9,7 @@ import { UserService } from '../user/user.service';
 import { MulterFile } from './interfaces/multer-file.interface';
 import { SearchCvDto } from './dto/search-cv.dto';
 
+
 @Injectable()
 export class CvService {
   constructor(
@@ -19,8 +20,21 @@ export class CvService {
   }
 
   async create(createCvDto: CreateCvDto , user: UserEntity) : Promise<CvEntity> {
-    const newCv = this.cvRepository.create({ ...createCvDto, user: user });
-    return this.cvRepository.save(newCv);
+
+    try{
+      const newCv = this.cvRepository.create({ ...createCvDto, user: user });
+      return await  this.cvRepository.save(newCv);
+    }catch (e) {
+      throw new HttpException('tous les champs firstname age cin job path sont obligatoirs', HttpStatus.BAD_REQUEST);
+    }
+
+
+  }
+  async findRelatedCvs(user: UserEntity): Promise<CvEntity[]> {
+    if(user.role === 'admin'){
+      return await this.cvRepository.find();
+    }
+    return await this.cvRepository.find({ where: { user } });
   }
   
   async associateFileWithCv(cvId: number, file: MulterFile): Promise<CvEntity> {
@@ -31,6 +45,8 @@ export class CvService {
    cv.path = file.path; 
    return this.cvRepository.save(cv);
   }
+
+
 
   async findAll() : Promise<CvEntity[]>{
     return await this.cvRepository.find();
@@ -46,10 +62,16 @@ export class CvService {
  
 
 
-  async  findOne(id: number) : Promise<CvEntity> {
+  async findOneById({ id, user }: { id: number, user: any }): Promise<CvEntity> {
     const cv = await this.cvRepository.findOne({ where: { id } });
-    return await cv;
-
+    if (!cv) {
+      throw new NotFoundException(`Le CV d'ID ${id} n'existe pas`);
+    }
+    if ( this.isOwnerOrAdmin(cv, user)) {
+      return cv;
+    } else {
+      throw new UnauthorizedException("Vous n'êtes pas autorisé à accéder à ce CV");
+    }
   }
 
   async update(id: number, cvDto: UpdateCvDto, user: UserEntity ): Promise<CvEntity> {
@@ -57,28 +79,37 @@ export class CvService {
   if (!existingCv) {
     throw new NotFoundException(`Le CV d'id ${id} n'existe pas`);
   }
-
-  if (cvDto.name !== undefined) {
+  if(this.isOwnerOrAdmin(existingCv, user) === false) {
+    throw new UnauthorizedException("Vous n'êtes pas autorisé à modifier ce CV");
+  }
+  if (cvDto.name ) {
     existingCv.name = cvDto.name;
   }
-  if (cvDto.age !== undefined) {
+
+  if (cvDto.age ) {
     existingCv.age = cvDto.age;
   }
-  if (cvDto.job !== undefined) {
+  if (cvDto.job ) {
     existingCv.job = cvDto.job;
   }
-  if (cvDto.path !== undefined) {
+  if (cvDto.path ) {
     existingCv.path = cvDto.path;
   }
-  if (user !== undefined) {
-    existingCv.user = user;
-  }
+
 
   const updatedCv = await this.cvRepository.save(existingCv);
 
   return updatedCv;
 }
-  async remove(id: number) {
+  async remove(id: number,user:UserEntity): Promise<any> {
+    const existingCv = await this.cvRepository.findOne({ where: { id } });
+
+    if (!existingCv) {
+      throw new NotFoundException(`Le CV d'id ${id} n'existe pas`);
+    }
+    if(this.isOwnerOrAdmin(existingCv, user) === false) {
+      throw new UnauthorizedException("Vous n'êtes pas autorisé à supprimer ce CV");
+    }
     return await this.cvRepository.delete(id);  }
 
 
@@ -97,21 +128,8 @@ export class CvService {
       }
     }
 
-async searchCvs(searchDto: SearchCvDto): Promise<CvEntity[]> {
-  const { age, criteria } = searchDto;
+    isOwnerOrAdmin(cv:CvEntity, user: UserEntity): boolean {
 
-  // Build query conditions based on provided search criteria
-  const queryBuilder = this.cvRepository.createQueryBuilder('cv');
-  if (age) {
-      queryBuilder.orWhere('cv.age = :age', { age });
-  }
-  if (criteria) {
-      queryBuilder.orWhere('cv.name LIKE :criteria', { criteria: `%${criteria}%` })
-          .orWhere('cv.firstname LIKE :criteria', { criteria: `%${criteria}%` })
-          .orWhere('cv.job LIKE :criteria', { criteria: `%${criteria}%` });
-  }
-
-  // Execute the query and return the result
-  return queryBuilder.getMany();
-}
+    return cv.user.id === user.id || user.role === 'admin';
+   }
 }
